moved progress bar to new helper fixed recursive logic for file copy
